+++ src/beamcodes/bcotable_risrc.txt
