+++ src/state_functions/density.py
